+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Reload AdPanel asynchronously
+- Ad server's event log
 ### Changed
 - AdServer configuration client uses API in camel case